requirements.txt
# Core data analysis
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
numpy>=1.21.0

# Web scraping and browser automation
playwright>=1.40.0
pytrends>=4.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.0.0

# Data validation
pydantic>=2.4.0
marshmallow>=3.20.0

# Rate limiting and retries
tenacity>=8.2.0
ratelimit>=2.2.1

# Progress tracking
tqdm>=4.66.0

# Date and time handling
python-dateutil>=2.8.2
pytz>=2023.3

# Optional: Advanced features
# google-trends-api>=0.1.0  # Alternative to pytrends
# aiohttp>=3.8.0  # For async scraping
# asyncio>=3.4.3  # For async support 